step_service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-dotenv
requests
pycryptodome
qrcode[pil]
ddddocr
Pillow
tls_client
